split file redux for search form
