+++ src/actions/searchAction.js
@@ -1,0 +1,6 @@
+export const action_search = (value) => {
+  return {
+    type: 'SEARCH_KEYWORD',
+    payload: value
+  }
+}
