+++ src/store.js
@@ -1,26 +1,10 @@
-import { createStore } from 'redux'
+import { createStore, combineReducers } from 'redux'
+import search from './reducers/searchReducer'
 
-const stateDefault = {
-  keyword: 'kosong'
-}
+let appReducers = combineReducers({
+  search
+})
 
-const reducer = (state = stateDefault, action) => {
-  switch (action.type) {
-    case 'SEARCH_KEYWORD': 
-      console.log('masuk 1')
-      state.keyword = action.payload
-      return state
-    default :
-      console.log('masuk default')
-      return state
-  }
-}
-
-export let action_search = {
-  type: 'SEARCH_KEYWORD',
-  payload: 'contoh'
-}
-
-const store = createStore(reducer)
+const store = createStore(appReducers)
 
 export default store
